base-research-stub/requirements.txt
web3>=6.0.0
python-dotenv
